Renamed NoCache to DummyCache
Turns out that we need to allocate a cache line so it could be
write back to lower level caches. If we do not allocate, the data
will be responded right to the sequencer and not be evicted.

Here, we use the smallest amount of cache that works with the simulator.
I.e., 2-way associative 256-byte cache, which can hold 4 cache lines.

Signed-off-by: Hoa Nguyen <hn@hnpl.org>